added new page layout
